refactor(storage): some refactor to unify hummock snapshot read
